+++ vm_thread.h
@@ -159,7 +159,7 @@ public:
   vm_state &process() { return _process; }
   vm_state const &process() const { return _process; }
 
-  int thread_index() const;
+  int64_t thread_index() const;
 
   vm_value return_value() const { return rp(); }
 };
